# [Core] Optimize interceptor runtime by leveraging compiled expressions.