--- a/src/SlimMessageBus.Host/AssemblyInfo.cs
+++ b/src/SlimMessageBus.Host/AssemblyInfo.cs
@@ -1,0 +1,20 @@
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// In SDK-style projects such as this one, several assembly attributes that were historically
+// defined in this file are now automatically added during build and populated with
+// values defined in project properties. For details of which attributes are included
+// and how to customise this process see: https://aka.ms/assembly-info-properties
+
+
+// Setting ComVisible to false makes the types in this assembly not visible to COM
+// components.  If you need to access a type in this assembly from COM, set the ComVisible
+// attribute to true on that type.
+
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM.
+
+[assembly: Guid("fa1c3d02-fc8e-41b9-bdbc-859a2cbda6cc")]
+
+[assembly: InternalsVisibleTo("SlimMessageBus.Host.Test")]