--- a/src/SlimMessageBus.Host/Collections/GenericTypeResolveCache.cs
+++ b/src/SlimMessageBus.Host/Collections/GenericTypeResolveCache.cs
@@ -1,0 +1,7 @@
+﻿namespace SlimMessageBus.Host.Collections;
+
+internal class GenericTypeResolveCache
+{
+    public bool IsEmpty { get; set; }
+    public bool IsSortRequred { get; set; }
+}