--- a/src/SlimMessageBus.Host/Collections/IReadOnlyCache.cs
+++ b/src/SlimMessageBus.Host/Collections/IReadOnlyCache.cs
@@ -1,0 +1,6 @@
+﻿namespace SlimMessageBus.Host.Collections;
+
+public interface IReadOnlyCache<TKey, TValue>
+{
+    TValue this[TKey key] { get; }
+}