--- a/src/SlimMessageBus.Host/Collections/IRuntimeTypeCache.cs
+++ b/src/SlimMessageBus.Host/Collections/IRuntimeTypeCache.cs
@@ -3,4 +3,9 @@
 public interface IRuntimeTypeCache
 {
     bool IsAssignableFrom(Type from, Type to);
+    TaskOfTypeCache GetTaskOfType(Type type);
+    /// <summary>
+    /// Cache for generic methods that match this signature <see cref="Func{TResult}"/>.
+    /// </summary>
+    IReadOnlyCache<(Type ClassType, string MethodName, Type GenericArgument), Func<object, object>> GenericMethod { get; }
 }