--- a/src/SlimMessageBus.Host/Producer/PublishContext.cs
+++ b/src/SlimMessageBus.Host/Producer/PublishContext.cs
@@ -1,0 +1,5 @@
+﻿namespace SlimMessageBus.Host;
+
+public class PublishContext : ProducerContext
+{
+}