--- a/src/SlimMessageBus.Host/Producer/SendContext.cs
+++ b/src/SlimMessageBus.Host/Producer/SendContext.cs
@@ -1,0 +1,8 @@
+﻿namespace SlimMessageBus.Host;
+
+public class SendContext : ProducerContext
+{
+    public DateTimeOffset Created { get; set; }
+    public DateTimeOffset Expires { get; set; }
+    public string RequestId { get; set; }
+}