… docs/intro.md
… docs/provider_memory.md
… src/SlimMessageBus.Host/Collections/GenericInterfaceTypeCache.cs
… src/SlimMessageBus.Host/Collections/GenericInterfaceTypeCache2.cs
… src/SlimMessageBus.Host/Collections/RuntimeTypeCache.cs
… src/SlimMessageBus.Host/Config/ConsumerSettings.cs
… src/SlimMessageBus.Host/Config/Fluent/AbstractConsumerBuilder.cs
… src/SlimMessageBus.Host/Config/Fluent/HandlerBuilder.cs
… src/SlimMessageBus.Host/Consumer/InterceptorPipelines/ConsumerInterceptorPipeline.cs
… src/SlimMessageBus.Host/Consumer/MessageProcessors/MessageHandler.cs
… src/SlimMessageBus.Host/Helpers/ReflectionUtils.cs
… src/SlimMessageBus.Host/MessageBusBase.cs
… src/SlimMessageBus.Host/Producer/InterceptorPipelines/ProducerInterceptorPipeline.cs
… src/SlimMessageBus.Host/Producer/InterceptorPipelines/PublishInterceptorPipeline.cs
… src/SlimMessageBus.Host/Producer/InterceptorPipelines/SendInterceptorPipeline.cs
… src/SlimMessageBus/Exceptions/ConsumerMessageBusException.cs
… src/SlimMessageBus/Exceptions/ProducerMessageBusException.cs
… src/SlimMessageBus/Exceptions/PublishMessageBusException.cs
… src/SlimMessageBus/Exceptions/RequestHandlerFaultedMessageBusException.cs
… src/SlimMessageBus/Exceptions/SendMessageBusException.cs
… src/SlimMessageBus/IPublishBus.cs
… src/SlimMessageBus/RequestResponse/IRequestResponseBus.cs
… src/Tests/SlimMessageBus.Host.Benchmark/ConsumerCallBenchmark.cs
… src/Tests/SlimMessageBus.Host.FluentValidation.Test/ConsumerValidationInterceptorTests.cs
… src/Tests/SlimMessageBus.Host.Integration.Test/HybridTests.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/AbstractMemoryBenchmark.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/Program.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/Properties/launchSettings.json
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/PubSubBenchmark.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/ReqRespBenchmark.cs